requirements.txt
Flask==2.0.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.2.2
Pillow==10.3.0
boto3